Refactor code to minimize imports for completion
Also update benchmark scripts.
