--- benchmarks/2_import_dit.py
@@ -1,1 +1,0 @@
-from dit import dit
