--- benchmarks/3_instantiate_dit_class.py
@@ -1,3 +1,0 @@
-from dit.dit import Dit
-
-dit = Dit()
