--- benchmarks/4_help_command.py
@@ -1,5 +1,0 @@
-from dit.dit import Dit
-
-dit = Dit()
-
-dit.interpret(['--help'])
