--- benchmarks/5_status_command.py
@@ -1,5 +1,0 @@
-from dit.dit import Dit
-
-dit = Dit()
-
-dit.interpret(['status'])
