+++ benchmarks/completion_command.py
@@ -1,0 +1,3 @@
+from dit.completion import interpret
+
+interpret(['1', 'dit'])
