+++ benchmarks/completion_index.py
@@ -1,0 +1,3 @@
+from dit.completion import interpret
+
+interpret(['4', 'dit', '-d', '../tests/ditdir', 'w'])
