+++ benchmarks/dit_help_command.py
@@ -1,0 +1,3 @@
+from dit.dit import interpret
+
+interpret(['--help'])
